feat: work around git-fetch flakiness
For some reason, git operations are not reliable.

Sometimes this is a network-related error (e.g. TLS-related failures).

Other times we've seen very odd behaviour around temporary
directories.

This latter may be related to (hopefully short-lived) problems running
Docker on Windows, but for now, exponential backoff combined with
explicity recreating the temporary directory in which the repo is
stored each time seems to help.