fix: Avoid including token in repo URL
Previously we tried to redact the token from anywhere it might leak to,
but this is fiddly and error prone. Git doesn't directly support
supplying the token in an environment variable, but it does allow us to
specify a custom "askpass" executable which can grab the token from the
environment for us.