feat: support external stata licence configuration
A stata licence can be configured by setting a STATA_LICENCE environment
variable, which will be needed for GHA and backend server usage.

For local runs, we add a best-effort convienience method - checking out a
private repo containing the opensafely license using the current user's git
credentials. If the user has access to the repo (either an opensafely member or
external collaborator), everything should just work.  If not, they'll get an
error message telling them to set the environment variable for their stata licence.

The goal of this is so we can make the stata docker image public, and avoid the
need for any alternative credentials for users - we just need github.

Note: the stata-mp image does not yet require a license - that change will land
once this one has.