feat: Add command for retrying failed job
This only applies to (and only works for) jobs which failed with an
Internal Error during the "finalise" step where we copy output files,
dump logs, update the manifest etc. When this happens we don't
automatically clean up the job's container and volume giving us an
opportunity to fix the bug and try again without having to re-run the
entire job.