fix: Fix passing of git auth tokens on non-Linux systems
The previous approach using GIT_ASKPASS worked fine on Linux but broke
mysteriously on macOS and didn't work on at all on Windows due to lack
of shebang support.

This restores our original approaching of just embedding the token in
the URL, which does require some redaction of exception messages but is
overall simpler.