fix: Work around Windows permissions issues
This is a (nasty) workaround for the permissions issues we hit when
switching between running the job-runner inside Docker and running it
natively on Windows. The issue is that the Docker process creates files
which the Windows native process then doesn't have permission to delete
or replace. We work around this here by using Docker to delete the
offending files for us.

This behaviour is enabled by setting the `ENABLE_PERMISSIONS_WORKAROUND`
environment variable to a non-empty value.

Note for the potentially confused: Windows permissions work nothing like
POSIX permissions. We can create new files in directories created by
Docker, we just can't modify or delete existing files.