feat: refactor logging setup and formatting
Move the formatting logic into a filter rather than the formatter. This avoids
the need for parameterising the formatter, it can be customised just using a
regular format string, which I then used in the 2 places that need it
(local_run and threaded service).

Additionally, changed the log format to put various tags at the end of the
output. This pushes line noise out to the left, but still allows us to filter,
e.g:

    grep project=foo log.txt | grep status=bar