fix: make windows service use a bash runner
This changes the nssm config to launch a bash script that sources .env
and then execs to python.

This means we do not need to do any env parsing in python, which is
good.

Also, we drop the OPENSAFELY_ prefix used by the current .env and
docker-compose tooling, just using the actual environment variables
directly.