fix: Generate consistent job IDs
Rather than randomly generating a job ID each time, we now use the hash
of the job request ID and the action name – which together are
guaranteed unique. This means even if we lose the entire job-runner
database while jobs are running we should be able to gracefully
recover because the same set of jobs with the same IDs should be
automatically recreated by the sync service.

Losing state like this still won't be _entirely_ unproblematic: in
particular, jobs which have already completed but are part of a job
request which is still active won't have their statuses correctly
reported. But it will at least avoid the problem of orphan jobs which
require a manual cleanup.