chore: Remove config value from wrapper script
This value has now been added to the environment file where it belongs.
Because the docker-compose file uses variables with an `OPENSAFELY_`
prefix we can set the config value for the native runner but not the
docker version. But it wouldn't actually be a problem to have it enabled
in both environments anyway.