Don't produce traceback for API errors
This is just useless noise in the logs, especially as we tend to get a
big batch of these together when there are API problems.

We also increase the time we wait after an error before retrying. In
future we can consider exponential backoff but this will do for now.