fix: Allow cohortextractor output in multiple dirs
Sometimes study authors want to create ancillary files alongside the
cohortextractor output. For instance, when study definitions or measures
are dynamically generated we may want to create a file recording the
parameters of the study/measure for use in later processing. Often it's
convenient to store this in a different directory from the standard
chortextractor output files.

The job-runner currently attempts to determine the expected directory
for cohortextractor output files by looking at the specified output
patterns and then appends this as an `--output-dir` argument to the
cohortextractor invocation. However if multiple output directories are
used in can't do this correctly and so throws an error.

This PR stops the error being thrown if an `--output-dir` is already
explicitly specified as one the cohortextractor arguments. The
assumption being that the user knows what they're doing in this case.

Longer term we should stop trying to automatically set `--output-dir`
and instead give the user a helpful error message if we can see that
their output patterns are never going to match.