feat: Pass `OPENSAFELY_BACKEND` env var to each job
This allows analysis scripts and study definitions to customise
themselves depending on which backend they are running under. This is
absolutely not the best long term solution to the cross-backend problem
but it's a simple and relatively non-invasive solution to unblock us for
now. We're deliberately not documenting this and only using it
internally at first.