feat: add PRESTO_TLS_* configs
PRESTO_TLS_KEY_PATH and PRESTO_TLS_CERT_PATH contain paths to the
relevant files, and we load them into PRESTO_TLS_KEY and PRESTO_TLS_CERT
respectively. If present, these values are passed to the cohortextractor
invocation.

This involved some error handing with in config.py that I wanted to
test, so I wrote a stupid subprocess test harness to do so.