feat: Support config file giving job "resource weightings"
This gives us a very crude mechanism for controlling how many of what
type of jobs are running together.

The config file itself should be placed at `job-resource-weights.ini`
in the job-runner working directory.

It is structured as:
```ini

[some-workspace-name]
my-ram-hungry-action = 4
other-actions.* = 1.5

[other-workspace-name]
...
```

Any jobs in the specified workspace will have their action names matched
against the regex patterns specified in the config file and will be
assigned the weight of the first matching pattern. All other jobs are
assigned a weight of 1.