Add basic ANSI formatting to output inside Github
The addition of the "expandable content" arrow in the Github output ends
up de-emphasising the `=> ` marker we were using to indicate action
names with the result that the whole output become visually much harder
to parse. As Github support ANSI escapes we can use these to bold the
action names and restore visual structure. We also take this opportunity
to de-emphasise the "(click here)" text.

Longer term we could make much better use of colour and formatting in
the output here, both within Github Actions and on the terminal more
generally. But that will require using a proper, well-tested library to
ensure we behave correctly on all the various platforms we need to
support.