test: Test filesystem for case sensitivity
Previously, we tested the operating system for case sensitivity.
However, this fails on some macOS installs because whilst macOS is
POSIX-compatible, by default APFS (Apple File System) is case insensitive.
Python's standard library doesn't have a function for testing the
filesystem; seemingly the best way is to write/read a temporary file.
Finally, because this function is specific to the test, rather than the
application code, it's next to the test.