feat: Handle reusable actions
We handle a reusable action by rewriting the action specification before
it's saved, to take advantage of existing functionality that fetches a
Git repository and copies it to a volume. We introduce a switch when
running a job, to ensure that reusable actions are fetched when running
in local run mode.

Closes #227