feat: Use timeouts when interacting with Docker
Shelling out to the Docker client on Windows sometimes hangs
indefinitely (for no obvious reason that we can determine). To prevent
this locking up the entire process we use a large-ish default timeout on
almost all calls to Docker.

The exceptions are copying files and pulling images which we expect to
sometimes take a long time. Pulling images is only done locally at the
moment in any case so we don't need to worry too much about that.
There's no simple way I can think of to distinguish a slow copy of a
large file from a broken call to Docker, so we'll just have to live with
that risk for now.