fix: Fix non-deterministic integration test
The addition of the `generate_cohort_with_dummy_data` action made the
dependency graph no longer have a single root so that when we started a
single job it was non-deterministic which one was started.

This PR pulls that actions out into a seperate job request. It also
slightly reorganises some of the test for (hopefully) added clarity.