fix: Change the default docker registry to be the proxy.
This avoids the need to do the tag dance we've been doing in prod, hence
removing scripts/update-docker-image and updating the playbook.

We do not want to use the proxy in local_run though, so we override it
to ghcr.io there, as well as the GIT_PROXY_DOMAIN, so that users are
hitting the services directly and not using our proxies.