feat: Pull docker images if they are not present
Previously, we required an operator to pull them before anything would
be able to run. This made set up and testing of jobrunner backends quite
slow. So we update on first use, but do not subsequently update, that is
still an operator action.

This meant removing the special case docker pull code in local_run, as
that was no longer needed. This provided an opportunity to clean up the
old stata docker workaround used in the old GHA scripts. All repos have
been updated to new action for a while now, so we don't need it any
more.