refactor: Refactor job creation to remove db access
This refactors the job creation and dependency resolution logic so it is
no longer intertwined with database access and so it no longer needs to
run within a transaction. As well as making the code clearer (I think)
this means we can now do expensive operations like pulling from Docker
or fetching from Github after we've worked out what we're going to run
but before we written anything to the database or opened a transaction.