refactor: Get action states from disk in single batch
This refactors the job creation process to do all disk access upfront in
a single batch, as we now already do for database access.

This means that the bulk of the job creation logic is now in a pure
function which takes a list of current jobs and returns a list of new
jobs.

As well as being a general structural improvement this paves the way for
support something other than local disk as the primary data store.