feat: Inital implementation of the spike API
This adds an EXECUTION_API feature flag, which switches the main run
loop to us alternate version of `handle_pending_jobs` and
`handle_running_jobs`, which use the API.

It adds unit tests for those new functions, using the fact we now have
an API! Fairly basic branch coverage to start with, but more to add.

Things this branch does not do:

 - implement selection and loading of the API implementation
 - handle manifest state

The spike code is mostly unaltered, except for moving the `set_message`
into `sync_job_status`, and adding an additional `update(job)` call.