Implement the current manifest logic in the initial API
This is a bit messy, but should hopefully be short lived. I tried to
reuse the various functions from `manage_jobs`, but that meant a bit of
fakery.

It also meant I need to properly sort out the factories for Job and
JobRequest, as a proper JobRequest is required for the manifest. I also
had to have actual directories on disk to work.

So the reaching around is a bit messy, as I guess it should be.