chore: Avoid setting the private repo token to an invalid value
The tests are set up to skip tests that need the token if it's not
set. Using the sample dotenv file as-is caused it to be set to an
invalid value, so the tests ran but failed. This change means that we
can successfully use the sample dotenv as is.