refactor: Switch to picking fields to update by exclusion rather than inclusion
It's a little difficult to keep track of which fields need persisting
when and missing a field could cause hard-to-understand bugs. The
motivation for just persisting specific fields was to avoid one thread
overwriting data that the other thread had written. In fact there is
only one case where this could happen -- when the sync thread marks a
job as cancelled and the run thread could overwrite it; so it's simpler
to just exclude that field from the updates.

(Note that the sync thread calls `update_where()` to just update the
cancelled flag rather than persisting the entire job, so there is no
risk of it overwriting the run thread's data.)