fix: Cancelled jobs should not modify workspace
At the moment, cancelled jobs are treated like failed jobs with the
consequence that they end up removing previous outputs from the same
action. We want them to act like jobs which were never really run.
However, for debugging purposes we still want to keep their logs and
metadata around. We do this by "part finalizing" them. This is not a
nice way of doing this, but this issue needs fixing now.

Fixes #246