Update tests for new split API
This involved some improvements to the StubJobAPI to make it suited

There may be more tests yet to add, but the basic cases seem covered.

There's a slight wrinkle in that the tests rely on the fact that the
StubJobAPI tracks the state of jobs. Not a problem, but smells a bit
like a layering violation.