feat: local docker JobAPI implementation
Adds a new EXECUTOR config to select backend, in
`dotted.module.path:class` format, after the common wsgi app notation.
Defaults to `jobrunner.executors.local:LocalDockerJobAPI`.

Attempted to split the docker work of old/new by using different
container and volume names.

Implementation tries to re-use as much of the `manage_jobs` functions as
possible, only re-writing the few parts that need it. This is a little
ugly, as the old way calls those methods with Job, where as the executor
calls them with a JobDefinition. This works because the attribute names
match, but is a little ugly. It may be worth duplicating all functions
into the local docker module as per Ben's spike.

Initial test coverage is mainly the integration test, which now runs
unmodified against both methods, and also some initial unit style tests
for individual stages (more to follow).