fix: List all available actions if no actions supplied
This was how the local_run CLI used to behave before it was accidentally
removed in c5625f50. Given that this is behavior we only need locally,
and given that we now have a seperate "create jobs" function for
local_run we can pull some of this code out of the main `project.py`
file.