Support local run the new local docker API
Given than most of the complexity of local run is on the scheduling
side, then this was actually remarkably simple.

I just had to copy the support for local git repo usage, and get the
local run tests to run with and without the API,

I also added docker_cleanup fixture, as the relying on the self-clean-up
abilities of local run meant that if there was an error, it would leave
docker containers behind, which would screw up subsequent test runs.