feat: Implement delete_files to clean up renamed files
This merges the WorkspaceAPI intro JobAPI, in order to make it simpler
to instantiate a single API instance.

The previous code used inodes to handle case issues, which doesn't work
across an API barrier, so this uses case insenstive comparison. This
could potentially leave files that just change case around in long term
storage, but this seemed a rare and acceptable failure mode.