ci: fix local docker dev setup to not use SSH.
Using SSH by default was slower, and error prone on different
developer's environments (e.g. they had to have SSH set up).

Instead, we default matching the docker GIDs so we can mount the socket.
This avoids the need for SSH in prod and dev

SSH is still supported however, as it is possibly useful in prod in the
longer term.