fix: run management containers for the duration.
Previously, they were created but not running. This meant that a well
intentioned but naive `docker volume prune` or `docker system prune`
would remove the container if the job was in the PREPARING or EXECUTING
states, which caused jobs to fail.

Keeping them running for the duration doesn't cost us any CPU, and only
~1MB of RAM, so seems reasonable to protect against accidental deletion
of the container or its volume.

Also, add labels to volume and management container as well as job
container, for easier debugging