fix: Be robust against `getuser()` failing on Windows
If `getpass.getuser()` can't find the current username from the
environment it falls back to importing `pwd`, which doesn't exist on
Windows. In addition it may fail in various other ways. See:
https://bugs.python.org/issue32731

Knowing the current username isn't crticial so we shouldn't fail if this
happens, just set the username to "unknown".