feat: add databuilder to integration tests
databuilder isn't currently in a place to support many of the columns
used by cohortextractor in the integration test so we need separate
dummy data and can't rely on the same actions.  For the same reason
test_integration has been broken down into two tests with the
expectation that it will be combined as databuilder adds the necessary
features.  The structure of test_integration has been updated so that
it's ready for parametrization in the future and to try and catch
inconsistencies when adding databuilder.