fix: generate requirements without hashes
Hashes mode causes pip install to fail when extra dependencies are
found.  This is not ideal when targetting multiple OSes when packages
have OS-optional extra dependencies (eg pytest needs atomicwrites on
win32 as of writing this).  Since we're not currently using pip with
hashes for this repo this unblocks using the just stack.