fix: atomic file writing and not following symlinks
We previously changed `docker.copy_from_volume()` to write to
a temporary file and rename, to gain basic atomic write behaviour.

This change generalises this and extends it to the non-docker
`copy_file()` method.

It also means we do not follow symlinks when writing to a file, as
`shutil.copy()` and friends do. This is useful in migrating between file
systems