Add Grype and Syft to workflows
See opensafely-core/security#1 for more context.

* Grype is a security scanner.
* Syft creates a software bill of materials (SBOM).

To add the software bill of materials, a release is created from the tag
in the `build_and_publish` workflow.

Reusable workflows are used here to try and avoid duplication. In
future, we could use reusable workflows or composite actions to share
the same steps across different repositories. For now, the proposal is
to just add to this repository as a trial.