fix: move the mode flag checking to start of job handling.
Previously, we cached the values of mode/paused flags at the top of the
loop. This meant if a flag is set part way through a loop, it would not
apply until the next loop started. Which could be a while.

So now we check the flag values for each job in the loop.