fix: send all current flag values to job-server
We now support a `mode` flag and a `pause`, as the two are independantly
enabled.

This change communicates all currently set flags to job-server, rather
than just the 'mode' flag. It encodes them as JSON string in a header,
which is a bit yuck, but this is a temporary thing until we have time
redesign the api.