Skip ActionSpecification tests on Windows
ActionSpecification objects are used to build up commands passed to
Docker.  We build up the commands with shlex.join() which uses
shlex.quote() under the hood to quote the given list items.  This is
only guaranteed to produce correct input on posix-compliant shells which
Windows typically does not have.  Since we're not running Windows inside
Docker we can skip running those tests on Windows.