fix: record timestamp for flag changes
Record when a flag was last set or unset. This will be helpful in
diagnosing issues, as it can help us see when things changed.

This did involve adding a new column to the Flag table, which we've not
got a mechanism to handle. But errors with the flag table are gracefully
handled, and its a simple manual ALTER TABLE to run on TPP (the only
place that currently has the flag table).