fix: record unmatched outputs in local executor
Previously, we had punted on collecting unmatched outputs when writing
the local executor, so this was missing from that code path.

In deleting the old code path, that functionality was lost.

This change adds it to the local executor flow. It seemed fairly simple
in the end, not sure why we did not do it the first time round.

There were not tests previously for this, so I added some. In doing so,
it seems the file timestamp logic had a bug, as it `docker cp` seemed to
preserve the timestamp of the src file, rather than add a new one. So
I changed to actually `touch` a new file rather than a copy.

The tests also exposed that docker's filesystem only supports 1s
precision on file timestamps, so the test had to sleep for 1 second.