chore: refactor local volume management
This is introduces a `volume_api` interface, in order to be able to
support different docker volume implmentations.

The only implementation is the current docker volume one, and it simply
wraps the various functions that were previously used. It does however
change the api to take a `job`, in order to:

a) make the the volume name opaque to the executor, in order that it can
   be a named volume or a host path.
b) `job` has all the information needed various volume api fuctions to
   work.

Doing this refactor has highlighted that the current implementation
places the responsibility of figuring out outputs and error on the
executor, when really this should be shared logic. Have written this up
as #425

This is related to fixing #407