chore: refactor db codes to be executor agnostic
These codes mean the same thing regardless of executor, so handle them
in the job loop rather than in the local executor.

This is also faster, as the tests do not need docker anymore

Also, this was discovered by trying to reorganise config.py in order to
group and documenet the general job-runner config from the local docker
executor config, so that reorganising of config.py is included.