feat: add a bind mount volume api
This volume api implementation uses host directories and bind
mounts them into the docker container.

This approach sould be more efficient than the docker volume api, which
has to `docker cp` everything onto the volume. This can just fast local
filesystem copy, using `os.sendfile` on linux.

This is tested on linux only, and had to be explicitly enabled via
config, and is indended to be used in production backends.