fix: handling initial executor state being ERROR
In our local docker implementation, this can't happen, as all
transitions to an ERROR state are driven synchronously by the main loop
calling the API.

However, in the Graphnet K8S environment, this happens asynchronously,
so the state of the job can be in ERROR at any point.

This change handles it being in the ERROR state before we even try to
transition it, which is the correct thing to do.

I had to tweak the StubExecutorAPI handle setting a message initially.

Fixes #440