fix: move docker image to run as root in prod
Running as non-root user by default was causing issues for graphnet, and
making it difficult to deploy to TPP, so regress to run running as root
for now.

Also, renamed the user inside the container as `opensafely` rather than
app users.

We still run as this user in local docker dev env, or else file
permissions get messy. But we do that in the docker-compose.yaml rather
than hardcoding in the dockerfile