feat: add fine grained status codes and track them in the db.
This means the a StatusCode is always required, and function signatures
have changed as a result.

Also, I noticed that that a few of our failure cases were being handled
in the handle_job(n) method, when really they should have been handled
by the generic exception handler to unify process, so fixed that in the
process.