feat: add opentelemetry tracing
This adds opentelemetry tracing to running jobs.

A new jobrunner/tracing.py module handles most of the work, and we use
it to trace state changes between the fine grained StatusCode states
added previously.  We store as many attributes on each span as possible.

This includes a migration to add somewhere to store a job's trace
context and last changed timestamp.

We also emit "marker spans" that indicate that a the job has entered
a certain state, but not yet exited it yet. This is intedended to be
useful for ongoing jobs.

Also, we add a final "RUN" span once the job has completed that serves
as coarser grained span to use for graphing whole job metics.