fix: make the root span duration last the whole job.
Also, fix the is_remote attribute on the propagated context, which had
not been caught before.

The motivation for this is that, while the short duration root span
works, both the OTel spec and the Honeycomb UI seem to make the
assumption that the root span lasts for the duration of all its
children.

Originally, we sent the root span immeadiately, so that it was present
in Honeycomb, and sent a suggorgate child span called RUN to track the
entire job's duration. However, not having a root span for an unfinished
trace does not seem to be a problem, and is a helpful indicator that the
job has not finished yet. So I think we can live with no root span in
Honeycomb until the job is finished.

In contrast, having a root span with a sensible duration has benefits in
the UI and querying, as Honeycomb can more easily highlight longer jobs.

However, it does create another small issue. We can not easily preserve
the python object representing the root span across job-runner restarts.
And the OTel API explicitly and deliberately does not provide a way to
create a span with a specific id.  So we have to cheat and set an
underscore prefix attribute to re-create an instance of the root span
when we need it.