fix: ensure trace attributes are valid types
We have seen some errors in production about some span attribute values
being None, but the logs do not say which attributes.

This change ensures all attributes are valid types. If they are not, it
coverts to string, and logs so we can find out where the bad attribute
is comming from.