fix: ensure mark_job_as_failed always marks trace as error
Previously, the user had to pass error=True (or error=exc) to get this.

To fix this, we make two changes

a) mark_job_as_failed require an explicit 'message'.
b) if `error` isn't explicity passed in, mark it as error anyway

Also, we fix up KILLED_BY_ADMIN

Fixes #497