Set test coverage target
This will cause the tests to fail if the code coverage is under this
percentage. It's initially been set to match the existing coverage amount,
although the hope is that this will increase over time.

Coverage is only run when running `just test` or `just docker/test`. This is
because other `just test*` options only run a subset of the tests, so we can't
get a useful result from the coverage tool. However, since `just docker/test`
is run in CI, it will always run for every code change.