fix: Tweak the sync protocol to make it more robust
This extends the notion of an active JobRequest to include not only all
those which job-server considers active but also any which job-runner
still considers active.

Under normal circumstances this change makes no difference as the former
is always a superset of the latter. But where a previously inactive job
has become active again (because of manual intervention by an admin)
then it's possible that job-runner can have active jobs which job-server
doesn't know about.

This is not a practice we want to encourage (and it will have other
unwanted effects e.g. with respect to metrics gathering) but
nevertheless it does happen sometimes and its better if we can robustly
handle this and always move towards stabilising on a consistent state
between job-runner and job-server.