chore: add loop tracing
This adds a new type of trace, a loop trace.

This will emit one trace per iteration of the main scheduling loop, with
one child span for each job it considers during that loop, with the
initial and final states recorded on each span.

The goal is to enable us to see the performance and behaviour of our
main scheduling loop, particular wrt to blocking transitions and other
slowness.