fix: add timestamp to JobStatus
This allows the backed to report precise timings of when a job finished
being PREPARED/EXECUTED/FINALIZED.

We plan to use this to trace state timings more accurately later.

In addtion, the need to check for the file timestamp exposed a latent
bug in `get_status()` where a volume could be created but not yet
finished being PREPARED. In practise, preparing is synchronous, so we
wouldn't see this bug, but the new implementation is a) more correct and
b) will work when we make preparing async and c) was basically free with
the need to check for a timestamp.