fix: add missings state trace spans with timings
Emit trace spans for the PREPARED/EXECUTED/FINALIZED states. Previously,
we'd only done if for the states the run loop was actively instigated.
These state changes happen asynchronously (or at least, can do) in the
executor, and were not being captured. This lead to in accuracies in
e.g. how long PREPARING steps ran for.

It uses the new JobStatus.timestamp field to get the timings.