ci: add some sanity assertions to tracing timestamp code.
We only want these to blow up in dev, not it prod.

So, we use the familiar `assert` syntax, but add a @warn_assertions
decorator to emit warnings for them. And we make warnings errors in the
pytest config.

This means if we hit these errors during tests, the test will fail. And
they will be emitted as warnings in prod too. But they won't raise in
production.