fix: add missings state trace spans with timings
Emit trace spans for the PREPARED/EXECUTED/FINALIZED states. Previously,
we'd only done it for the states which run loop actively instigated.
These state changes happen asynchronously (or at least, can do) in the
executor, and were not being captured. This leads to in accuracies in
e.g. how long PREPARING steps ran for.

It uses the new JobStatus.timestamp_ns field returned from the executor
to get the timings.