fix: add missings state trace spans with timings
Emit trace spans for the PREPARED/EXECUTED/FINALIZED states. Previously,
we'd only done it for the states which run loop actively instigated.
These state changes happen asynchronously (or at least, can do) in the
executor, and were not being captured. This leads to in accuracies in
e.g. how long PREPARING steps ran for.

It uses the new JobStatus.timestamp field returned from the executor to
get the timings.

There is a wrinkle - the max precision of the timestamp from the stat we
get from stat'ing the opensafely-timestamp file is truncated second.
This means we can actually go back in time, which causes an integer
overflow in honeycomb, and results in infeasible long spans.

To address this we a) add 1 to such timestamps, basically doing
`ceil()` rather than `floor()`, to ensure we don't go backwards b)
handle the case when the end_time < start_timne, by moving the end_time
to 1ms after the start_time, which seemed a safe thing to do.