fix: change JobStatus.timestamp to ns precision
Using seconds meant truncatiing to nearest second, which made it
possible for telemetry timestamps to move backwards in time which causes
all kinds of complexity.

This changes the semantics of JobStatus.timestamp to be integer
nanoseconds, which should mean it never goes backwards. To support this
for the PREPARED state, we write the ns timestamp into as the contents
of the .opensafely-timestamp, so we can read them out later.  For old
jobs, we fall back to using the filesystem metadata, but via an
alternative stat invocation from the origina, which gives us higher
resolution timestamps.

Also, reworked and tested the parsing of datestrings to handle ns values
and Z utc specficier.  This means there is no truncation at all.

Renamed the `touch_file/file_timestamp` functions to
write_timestamp/read_timestamp` to reflect their goal rather than
implementation.

This doesn't change the find new files logic, which still uses the
filesystem mtime metadata for finding new files, via `find -newer`