fix: don't error the job on ExecutorRetry
Instead, record that it happened, but don't fail the job. The idea is
that we can then be alerted to it happening, but when it's over, the job
continues.

Also, increase get_status timeout from 10s to 15s, we a) occasionally we
saw it in GHA CI, and b) having it slow down a bit when docker is
blocking is no bad thing