refactor: subsume set_state into set_code
This avoids scattering the state change code in multiple places, and
means it's more consistant. It also continues the direction of making
`StatusCode` the key thing we care about, and the coarser, user-focused
`State` subservient to it.

This will allow the next commit to not have to worry about setting
job.state.